chore: move BlockHeaderResponse to http package
